+++ swift_browser_ui/upload/api.py
@@ -109,6 +109,10 @@ async def handle_replicate_object(request: aiohttp.web.Request) -> aiohttp.web.R
         source_container,
     )
 
+    # Ensure that both containers exist
+    await replicator.a_ensure_container()
+    await replicator.a_ensure_container(segmented=True)
+
     asyncio.ensure_future(replicator.a_copy_object(source_object))
 
     return aiohttp.web.Response(status=202)